documentation comments and cleaning on execution
