+++ packages/rs-drive-abci/src/execution/mod.rs
@@ -1,24 +1,22 @@
-/// The block proposal
-pub mod block_proposal;
 /// Data triggers
 pub mod data_trigger;
 /// Engine module
 pub mod engine;
-/// An execution event
-pub mod execution_event;
 /// Fee pools module
 pub mod fee_pools;
-/// A clean version of the the requst to finalize a block
-pub mod finalize_block_cleaned_request;
 /// Helper methods
 pub mod helpers;
 /// Initialization
 pub mod initialization;
 /// Masternode Identities
 mod masternode_identities;
+/// Processor module
+pub mod processor;
 /// Protocol upgrade
 pub mod protocol_upgrade;
-/// Quorum methods
-pub mod quorum;
-/// Test quorum for mimic block execution
-pub mod test_quorum;
+/// Types needed in execution
+mod types;
+/// Validator set update module
+pub mod validator_set_update;
+
+pub use types::*;
